# cedar-openview-server-application/src/main/resources/assets/swagger-api/swagger.yaml
openapi: 3.0.1
info:
  description: |
    Better Metadata - Better Science

    The OpenView server hands out CEDAR content that has been marked open, to anyone, without
    credentials. It is what makes a published artifact citable: the artifact IRI resolves here for a
    reader who has no CEDAR account.

    An artifact or folder is served when it is open itself, or when a folder above it is. Anything
    else is refused, and nothing here writes.
  title: CEDAR OpenView Server API
  version: 2.9.8-SNAPSHOT
paths:
  /folders/{id}:
    get:
      description: "Return what an open folder holds, with the path back to the workspace\
        \ root. A folder is served when it is marked open, or when a folder above\
        \ it is. No credentials are involved: this server exists to hand out open\
        \ content anonymously."
      operationId: findFolder
      parameters:
      - description: Folder identifier.
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: "Comma-separated artifact types to include, in place of all of\
          \ them."
        in: query
        name: resource_types
        schema:
          type: string
      - description: "Which versions to include: `latest`, `latest-published`, `latest-draft`,\
          \ or `all`."
        in: query
        name: version
        schema:
          type: string
      - description: "Filter by publication status: `draft` or `published`."
        in: query
        name: publication_status
        schema:
          type: string
      - description: Comma-separated sort fields; a leading `-` reverses one.
        in: query
        name: sort
        schema:
          type: string
      - description: Maximum number of entries to return.
        in: query
        name: limit
        schema:
          type: integer
          format: int32
      - description: Number of entries to skip before the first one returned.
        in: query
        name: offset
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: The folder's contents and its path
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: "A paging, sort, or filter parameter is not valid"
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: "The folder exists but is not open, and neither is any folder\
            \ above it"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: No such folder
      summary: List the contents of an open folder
      tags:
      - Folders
  /template-elements/{id}:
    get:
      description: "Return a template element that is open to everyone. An artifact\
        \ is served when it is marked open, or when it sits under a folder that is.\
        \ No credentials are involved: this server exists to hand out open artifacts\
        \ anonymously, which is what makes a published CEDAR artifact citable. Mongo's\
        \ internal `_id` is removed before the artifact is returned."
      operationId: findTemplateElement
      parameters:
      - description: Artifact identifier. Either the bare identifier or the full IRI
          is accepted; a bare one is resolved to the IRI before lookup.
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The template element
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: "The template element exists but is not open, and neither is\
            \ any folder above it"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: No such template element
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: The artifact store could not be reached after the template
            element was found to be open
      summary: Get an open template element
      tags:
      - Template elements
  /template-fields/{id}:
    get:
      description: "Return a template field that is open to everyone. An artifact\
        \ is served when it is marked open, or when it sits under a folder that is.\
        \ No credentials are involved: this server exists to hand out open artifacts\
        \ anonymously, which is what makes a published CEDAR artifact citable. Mongo's\
        \ internal `_id` is removed before the artifact is returned."
      operationId: findTemplateField
      parameters:
      - description: Artifact identifier. Either the bare identifier or the full IRI
          is accepted; a bare one is resolved to the IRI before lookup.
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The template field
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: "The template field exists but is not open, and neither is\
            \ any folder above it"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: No such template field
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: The artifact store could not be reached after the template
            field was found to be open
      summary: Get an open template field
      tags:
      - Template fields
  /template-instances/{id}:
    get:
      description: "Return a template instance that is open to everyone. An artifact\
        \ is served when it is marked open, or when it sits under a folder that is.\
        \ No credentials are involved: this server exists to hand out open artifacts\
        \ anonymously, which is what makes a published CEDAR artifact citable. Mongo's\
        \ internal `_id` is removed before the artifact is returned."
      operationId: findTemplateInstance
      parameters:
      - description: Artifact identifier. Either the bare identifier or the full IRI
          is accepted; a bare one is resolved to the IRI before lookup.
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The template instance
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: "The template instance exists but is not open, and neither\
            \ is any folder above it"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: No such template instance
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: The artifact store could not be reached after the template
            instance was found to be open
      summary: Get an open template instance
      tags:
      - Template instances
  /templates/{id}:
    get:
      description: "Return a template that is open to everyone. An artifact is served\
        \ when it is marked open, or when it sits under a folder that is. No credentials\
        \ are involved: this server exists to hand out open artifacts anonymously,\
        \ which is what makes a published CEDAR artifact citable. Mongo's internal\
        \ `_id` is removed before the artifact is returned."
      operationId: findTemplate
      parameters:
      - description: Artifact identifier. Either the bare identifier or the full IRI
          is accepted; a bare one is resolved to the IRI before lookup.
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The template
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: "The template exists but is not open, and neither is any folder\
            \ above it"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: No such template
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: The artifact store could not be reached after the template
            was found to be open
      summary: Get an open template
      tags:
      - Templates
components:
  schemas:
    CedarError:
      type: object
      additionalProperties: true
      description: A CEDAR error response. Diagnostic fields are populated only when
        they are relevant to the failure.
      properties:
        entities:
          type: object
          additionalProperties:
            type: object
            description: Named domain entities associated with exception-mapped errors.
            nullable: true
          description: Named domain entities associated with exception-mapped errors.
          nullable: true
        errorId:
          type: string
          format: uuid
          description: Correlation identifier for an internal error recorded in server
            logs.
          nullable: true
        errorKey:
          type: string
          description: Stable machine-readable error identifier.
          enum:
          - templateElementNotCreated
          - templateElementNotFound
          - templateElementNotDeleted
          - templateElementNotUpdated
          - templateElementsNotListed
          - templateFieldNotCreated
          - templateFieldNotFound
          - templateFieldNotDeleted
          - templateFieldNotUpdated
          - templateFieldsNotListed
          - templateNotCreated
          - templateNotFound
          - templateNotDeleted
          - templateNotUpdated
          - templatesNotListed
          - templateInstanceNotCreated
          - templateInstanceNotFound
          - templateInstanceNotDeleted
          - templateInstanceNotUpdated
          - templateInstancesNotListed
          - noReadAccessToResource
          - noWriteAccessToResource
          - noReadAccessToFolder
          - noWriteAccessToFolder
          - noReadAccessToTemplate
          - noWriteAccessToTemplate
          - noReadAccessToTemplateElement
          - noWriteAccessToTemplateElement
          - noReadAccessToTemplateField
          - noWriteAccessToTemplateField
          - noReadAccessToTemplateInstance
          - noWriteAccessToTemplateInstance
          - noReadAccessToArtifact
          - noWriteAccessToArtifact
          - noWriteAccessToCategory
          - folderNotFound
          - artifactNotFound
          - artifactPreconditionRequired
          - artifactHasMovedOn
          - verbatimWriteRefused
          - nodeNotFound
          - unknownResourceType
          - missingParameter
          - invalidInput
          - sourceFolderNotFound
          - sourceResourceNotFound
          - targetFolderNotFound
          - nodeNotMoved
          - groupNotFound
          - userNotFound
          - uniqueConstraintCollision
          - invalidData
          - notAuthorized
          - permissionMissing
          - tokenInvalid
          - userInfoLoadByTokenFailed
          - userInfoLoadByApiKeyFailed
          - cedarUserNotFound
          - authorizationNotFound
          - permissionNotOwned
          - authorizationTypeUnknown
          - apiKeyNotFound
          - tokenMissing
          - tokenExpired
          - parentFolderNotSpecified
          - parentFolderSpecifiedTwice
          - pathNotNormalized
          - parentFolderNotFound
          - updateInvalidFolderName
          - createInvalidFolderName
          - nodeAlreadyPresent
          - folderNotCreated
          - missingNameAndDescription
          - missingData
          - folderNotDeleted
          - artifactNotDeleted
          - resourceNotCreated
          - invalidResourceType
          - invalidArtifactType
          - readOtherProfileForbidden
          - updateOtherProfileForbidden
          - folderCanNotBeDeleted
          - folderCanNotBeChanged
          - groupAlreadyPresent
          - groupCanBeModifiedOnlyByGroupAdmin
          - groupCanBeDeletedOnlyByGroupAdmin
          - groupUsersNotUpdated
          - specialGroupCanNotBeDeleted
          - folderPermissionsCanNotBeChanged
          - unknownInstanceOutputFormat
          - folderCopyNotAllowed
          - methodNotImplemented
          - upstreamServerError
          - nothingToDo
          - parentCategoryNotFound
          - categoryAlreadyPresent
          - categoryNotFound
          - categoryCanNotBeDeleted
          - rootCategoryCanNotBeDeleted
          - unableToAttachCategory
          - unableToDetachCategory
          - noCategoriesWereAttached
          - malformedJsonRequestBody
          - malformedSearchTerm
          - publishedArtifactCanNotBeChanged
          - publishedArtifactCanNotBeDeleted
          - versioningOnlyOnLatest
          - versioningOnlyByOwner
          - nonVersionedArtifactType
          - createDraftOnlyFromPublished
          - publishOnlyDraft
          - draftNotCreated
          - contentNotValid
          - doiNotSupportedByResourceType
          - doiCanNotBeSetForEmptyAtId
          - doiCanNotBeAltered
          - doiCanNotBeSet
          - doiAlreadyExists
          - dataCiteDOIDisabled
          - resourceNotFound
          - internalError
          nullable: true
        errorMessage:
          type: string
          description: Alias of message retained for compatibility.
          nullable: true
        errorReasonKey:
          type: string
          description: More specific machine-readable reason for the error.
          enum:
          - nonEmptyFolder
          - nonEmptyCategory
          - userHomeFolder
          - systemFolder
          - templateReferencedInInstances
          - validationError
          - continuationExpired
          nullable: true
        errorType:
          type: string
          description: Broad category of the error.
          enum:
          - notFound
          - invalidArgument
          - authorization
          - permission
          - server
          - validationError
          nullable: true
        message:
          type: string
          description: Human-readable error message.
          nullable: true
        objects:
          type: object
          additionalProperties:
            type: object
            description: Named structured values associated with the error.
          description: Named structured values associated with the error.
        operation:
          type: object
          additionalProperties:
            type: object
            description: Operation that failed. Its fields depend on the operation
              type.
            nullable: true
          description: Operation that failed. Its fields depend on the operation type.
          nullable: true
        parameters:
          type: object
          additionalProperties:
            type: object
            description: Named scalar values associated with the error.
          description: Named scalar values associated with the error.
        status:
          type: string
          description: Symbolic HTTP response status.
          enum:
          - BAD_REQUEST
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - METHOD_NOT_ALLOWED
          - NOT_ACCEPTABLE
          - CONFLICT
          - PRECONDITION_FAILED
          - UNSUPPORTED_MEDIA_TYPE
          - UNPROCESSABLE_ENTITY
          - PRECONDITION_REQUIRED
          - INTERNAL_SERVER_ERROR
          - NOT_IMPLEMENTED
          - BAD_GATEWAY
          - SERVICE_UNAVAILABLE
          - HTTP_VERSION_NOT_SUPPORTED
        statusCode:
          type: integer
          format: int32
          description: Numeric HTTP response status.
          maximum: 599
          minimum: 400
        suggestedAction:
          type: string
          description: Recovery action suggested to the client.
          enum:
          - none
          - requestRole
          - logout
          - logoutImmediately
          - provideAuthorizationHeader
          - refreshToken
          nullable: true
      required:
      - status
      - statusCode
  securitySchemes:
    api_key:
      description: "Your CEDAR API key, using the format 'apiKey <yourApiKey>'"
      in: header
      name: Authorization
      type: apiKey